feat/slide_card 작성중-#38
